Avoid reflection in AndroidCertificateChainCleaner (#5961)
* Avoid reflection in AndroidCertificateChainCleaner

* clean
